+++ packages/eslint-config-next/src/index.js
@@ -12,13 +12,10 @@ const isTSExist = fs.existsSync(join(process.cwd(), tsconfig))
 
 const tsconfigRootDir = process.cwd()
 
-const a11yOff = Object.keys(require('eslint-plugin-jsx-a11y').rules).reduce(
-  (acc, rule) => {
-    acc[`jsx-a11y/${rule}`] = 'off'
-    return acc
-  },
-  {}
-)
+const a11yOff = Object.keys(require('eslint-plugin-jsx-a11y').rules).reduce((acc, rule) => {
+  acc[`jsx-a11y/${rule}`] = 'off'
+  return acc
+}, {})
 
 module.exports = defineConfig({
   root: true,
@@ -37,8 +34,6 @@ module.exports = defineConfig({
     'airbnb-typescript',
     'airbnb/hooks',
     'plugin:@typescript-eslint/recommended',
-    'plugin:@typescript-eslint/recommended-type-checked',
-    'plugin:@typescript-eslint/stylistic-type-checked',
     'plugin:react-hooks/recommended',
     'plugin:import/recommended',
     'plugin:import/typescript',
@@ -46,26 +41,11 @@ module.exports = defineConfig({
     'plugin:import/warnings',
     'prettier'
   ],
-  plugins: [
-    '@typescript-eslint',
-    'react',
-    'simple-import-sort',
-    'import',
-    'unused-imports'
-  ],
+  plugins: ['@typescript-eslint', 'react', 'simple-import-sort', 'import', 'unused-imports'],
   settings: {
     'import/resolver': {
       node: {
-        extensions: [
-          '.js',
-          '.cjs',
-          '.mjs',
-          '.ts',
-          '.cts',
-          '.mts',
-          '.tsx',
-          '.d.ts'
-        ]
+        extensions: ['.js', '.cjs', '.mjs', '.ts', '.cts', '.mts', '.tsx', '.d.ts']
       }
     }
   },
@@ -112,13 +92,7 @@ module.exports = defineConfig({
       'warn',
       {
         props: true,
-        ignorePropertyModificationsFor: [
-          'target',
-          'descriptor',
-          'req',
-          'request',
-          'args'
-        ]
+        ignorePropertyModificationsFor: ['target', 'descriptor', 'req', 'request', 'args']
       }
     ], // 允许修改函数参数，但是会有警告