+++ packages/eslint-config-react/src/index.js
@@ -12,13 +12,10 @@ const isTSExist = fs.existsSync(join(process.cwd(), tsconfig))
 
 const tsconfigRootDir = process.cwd()
 
-const a11yOff = Object.keys(require('eslint-plugin-jsx-a11y').rules).reduce(
-  (acc, rule) => {
-    acc[`jsx-a11y/${rule}`] = 'off'
-    return acc
-  },
-  {}
-)
+const a11yOff = Object.keys(require('eslint-plugin-jsx-a11y').rules).reduce((acc, rule) => {
+  acc[`jsx-a11y/${rule}`] = 'off'
+  return acc
+}, {})
 
 module.exports = defineConfig({
   root: true,
@@ -36,8 +33,6 @@ module.exports = defineConfig({
     'airbnb-typescript',
     'airbnb/hooks',
     'plugin:@typescript-eslint/recommended',
-    'plugin:@typescript-eslint/recommended-type-checked',
-    'plugin:@typescript-eslint/stylistic-type-checked',
     'plugin:react-hooks/recommended',
     'plugin:@tanstack/eslint-plugin-query/recommended',
     'plugin:import/recommended',
@@ -57,16 +52,7 @@ module.exports = defineConfig({
   settings: {
     'import/resolver': {
       node: {
-        extensions: [
-          '.js',
-          '.cjs',
-          '.mjs',
-          '.ts',
-          '.cts',
-          '.mts',
-          '.tsx',
-          '.d.ts'
-        ]
+        extensions: ['.js', '.cjs', '.mjs', '.ts', '.cts', '.mts', '.tsx', '.d.ts']
       }
     }
   },
@@ -113,13 +99,7 @@ module.exports = defineConfig({
       'warn',
       {
         props: true,
-        ignorePropertyModificationsFor: [
-          'target',
-          'descriptor',
-          'req',
-          'request',
-          'args'
-        ]
+        ignorePropertyModificationsFor: ['target', 'descriptor', 'req', 'request', 'args']
       }
     ], // 允许修改函数参数，但是会有警告
 
@@ -191,10 +171,7 @@ module.exports = defineConfig({
     ],
 
     // react-refresh
-    'react-refresh/only-export-components': [
-      'warn',
-      { allowConstantExport: true }
-    ],
+    'react-refresh/only-export-components': ['warn', { allowConstantExport: true }],
 
     // jsx-a11y
     ...a11yOff, // 禁用所有 jsx-a11y 规则