+++ packages/eslint-config-next/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brucesong/eslint-config-next
 
+## 0.0.5
+
+### Patch Changes
+
+- update typescript eslint config
+
 ## 0.0.4
 
 ### Patch Changes